+++ backends/functional/test_generic.cc
@@ -18,7 +18,7 @@
  */
 
 #include "kernel/yosys.h"
-#include "kernel/functionalir.h"
+#include "kernel/functional.h"
 #include <random>
 
 USING_YOSYS_NAMESPACE
@@ -139,7 +139,7 @@ struct FunctionalTestGeneric : public Pass
 
 		for (auto module : design->selected_modules()) {
             log("Dumping module `%s'.\n", module->name.c_str());
-			auto fir = FunctionalIR::from_module(module);
+			auto fir = Functional::IR::from_module(module);
 			for(auto node : fir)
 				std::cout << RTLIL::unescape_id(node.name()) << " = " << node.to_string([](auto n) { return RTLIL::unescape_id(n.name()); }) << "\n";
 			for(auto [name, sort] : fir.outputs())
